Migrate [gv_entry_link] to the 2.0 shortcode class
See https://gravitykit.slack.com/archives/G59JBVCT0/p1756159115769399
