+++ future/includes/class-gv-core.php
@@ -171,10 +171,12 @@ final class Core {
 		require_once $this->plugin->dir( 'future/includes/class-gv-shortcode-gventry.php' );
 		require_once $this->plugin->dir( 'future/includes/class-gv-shortcode-gvfield.php' );
 		require_once $this->plugin->dir( 'future/includes/class-gv-shortcode-gvlogic.php' );
+		require_once $this->plugin->dir( 'future/includes/class-gv-shortcode-gv-entry-link.php' );
 		add_action( 'init', array( '\GV\Shortcodes\gravityview', 'add' ) );
 		add_action( 'init', array( '\GV\Shortcodes\gventry', 'add' ) );
 		add_action( 'init', array( '\GV\Shortcodes\gvfield', 'add' ) );
 		add_action( 'init', array( '\GV\Shortcodes\gvlogic', 'add' ) );
+		add_action( 'init', array( '\GV\Shortcodes\gv_entry_link', 'add' ) );
 
 		/** oEmbed */
 		require_once $this->plugin->dir( 'future/includes/class-gv-oembed.php' );
